+++ ui/shared/Web3ModalProvider.tsx
@@ -1,100 +1,24 @@
-import type { AppKitNetwork } from '@reown/appkit/networks';
-import { createAppKit, useAppKitTheme } from '@reown/appkit/react';
+import dynamic from 'next/dynamic';
 import React from 'react';
 import { WagmiProvider } from 'wagmi';
 
-import config from 'configs/app';
-import useIsMounted from 'lib/hooks/useIsMounted';
-import { chains } from 'lib/web3/chains';
 import wagmiConfig from 'lib/web3/wagmiConfig';
-import { useColorMode } from 'toolkit/chakra/color-mode';
-import colors from 'toolkit/theme/foundations/colors';
-import { BODY_TYPEFACE } from 'toolkit/theme/foundations/typography';
-import zIndex from 'toolkit/theme/foundations/zIndex';
 
-const feature = config.features.blockchainInteraction;
-
-const init = () => {
-  try {
-    if (!feature.isEnabled || !wagmiConfig.adapter) {
-      return;
-    }
-
-    createAppKit({
-      adapters: [ wagmiConfig.adapter ],
-      networks: chains as [AppKitNetwork, ...Array<AppKitNetwork>],
-      metadata: {
-        name: `${ config.chain.name } explorer`,
-        description: `${ config.chain.name } explorer`,
-        url: config.app.baseUrl,
-        icons: [ config.UI.navigation.icon.default ].filter(Boolean),
-      },
-      projectId: feature.walletConnect.projectId,
-      features: {
-        analytics: false,
-        email: false,
-        socials: [],
-        onramp: false,
-        swaps: false,
-      },
-      themeVariables: {
-        '--w3m-font-family': `${ BODY_TYPEFACE }, sans-serif`,
-        '--w3m-accent': colors.blue[600].value,
-        '--w3m-border-radius-master': '2px',
-        '--w3m-z-index': zIndex?.modal2?.value,
-      },
-      featuredWalletIds: feature.walletConnect.featuredWalletIds,
-      allowUnsupportedChain: true,
-    });
-  } catch (error) {}
-};
+const Web3ModalProviderClient = dynamic(() => import('./Web3ModalProviderClient'), {
+  ssr: false,
+});
 
 interface Props {
   children: React.ReactNode;
 }
 
-const DefaultProvider = ({ children }: Props) => {
+const Web3ModalProvider = ({ children }: Props) => {
   return (
     <WagmiProvider config={ wagmiConfig.config }>
-      { children }
-    </WagmiProvider>
-  );
-};
-
-const Web3ModalProviderClient = ({ children }: Props) => {
-  const { colorMode } = useColorMode();
-  const { setThemeMode } = useAppKitTheme();
-
-  React.useEffect(() => {
-    setThemeMode(colorMode ?? 'light');
-  }, [ colorMode, setThemeMode ]);
-
-  React.useEffect(() => {
-    init();
-  }, [ ]);
-
-  return (
-    <DefaultProvider>
-      { children }
-    </DefaultProvider>
-  );
-};
-
-const Web3ModalProvider = ({ children }: Props) => {
-  const isMounted = useIsMounted();
-
-  if (!feature.isEnabled || !isMounted) {
-    return (
-      <DefaultProvider>
+      <Web3ModalProviderClient>
         { children }
-      </DefaultProvider>
-    );
-  }
-
-  return (
-    <Web3ModalProviderClient>
-      { children }
-    </Web3ModalProviderClient>
+      </Web3ModalProviderClient>
+    </WagmiProvider>
   );
 };
